+++ src/main/kotlin/com/github/gllxl/imagepreview/JsLineMarkerContributor.kt
@@ -1,23 +1,28 @@
 package com.github.gllxl.imagepreview
 
-import com.intellij.execution.lineMarker.RunLineMarkerContributor
+import com.github.gllxl.imagepreview.ImageMapping.setLineMapping
 import com.intellij.lang.javascript.psi.JSVariable
+import com.intellij.psi.PsiDocumentManager
 import com.intellij.psi.PsiElement
+import com.intellij.refactoring.suggested.startOffset
 
 fun getJsVariableContent (element: PsiElement): String {
   return element.node.lastChildNode.text
 }
 
-class JsLineMarkerContributor : RunLineMarkerContributor() {
+class JsLineMarkerContributor : LineMakerContributor() {
   override fun getInfo(element: PsiElement): Info? {
 
     if (element is JSVariable) {
-      val firstChild = element.firstChild
-      if (!firstChild.isValid) {
+
+      if (!element.firstChild.isValid) {
         return null
       }
 
+      val document = PsiDocumentManager.getInstance(element.project).getDocument(element.containingFile) ?: return null
+      val lineNumber = document.getLineNumber(element.startOffset)
       val imgUrl = getJsVariableContent(element)
+      setLineMapping(element.containingFile.virtualFile, lineNumber, removeUrlQuotes(imgUrl))
 
       return getLineMaker(removeUrlQuotes(imgUrl))
     }