+++ CHANGELOG.md
@@ -2,6 +2,45 @@
 
 # image-preview Changelog
 
+## [0.1.7] - 2024-03-2
+
+### Chore
+-  Support 2024.1
+
+## [0.1.6] - 2023-07-11
+
+### Chore
+-  Remove plugin version limit
+
+## [0.1.5] - 2023-05-21
+
+### Added
+-  Added image size info
+
+## [0.1.4] - 2023-04-20
+
+### Fixed
+-  Fixed LinePainter wrong display
+
+## [0.1.3] - 2023-04-19
+
+### Added
+-  Added Line Painter ( show image size )
+-  Support 2023.1
+
+### Fixed
+- Fix the error caused by wrong URL
+
+## [0.1.2] - 2023-03-23
+
+### Added
+-  Added ImagePool to caching images
+
+## [0.1.1] - 2023-03-20
+
+### Fixed
+- fix popup position
+
 ## [0.1.0] - 2023-03-16
 
 ### Added